creating-aux-dataset-with-bbox-clipped-labels-mapped-to-image-space

+++ ecology_semantic_segmentation/dataset/utils.py
@@ -8,35 +8,27 @@ import traceback
 import cv2
 import numpy as np
 
-from skimage.metrics import structural_similarity as ssim
+# slightly slower than MSE - able to achieve results with MSE
+#from skimage.metrics import structural_similarity as ssim
 
+from .composite_bboxes import BBOX_ANNOTATION_FILES
+
+# Tried thresholding but found similarity based brute force bbox assignment to work better for "whole body" annotations
 WHITE = ((0, 0, 237), (181, 25, 255))
 BLACKGREY = ((0, 0, 0), (180, 255, 35))
 
-BBOX_ANNOTATION_FILES = [
-                "12-23-2019/original image/f92B.jpg",
-                "photos 1.30.2019/original image/f9aB.jpg",
-                "photos 1.30.2019/original image/f97B.jpg",
-                "12-23-2019/original image/f93B.jpg",
-                "photos 1.30.2019/original image/f96B.jpg",
-                "photos 1.30.2019/original image/f99B.jpg",
-                "photos 1.30.2019/original image/f25B.jpg", 
-                "photos 1.30.2019/original image/f6B.png", 
-                "photos 1.30.2019/original image/f15aB.jpg", 
-                "photos 1.30.2019/original image/f7B.png", 
-                "photos 1.30.2019/original image/f99B.jpg", 
-                "photos 1.30.2019/original image/f199B.jpg", 
-                #"photos 1.30.2019/original image/f45B.png",
-                #"photos 1.30.2019/original image/f34B.jpg",
-                #"12-23-2019/original image/f58B.jpg",
-                #"photos 1.30.2019/original image/f52B.jpg",
-                        ]
-
-ORIGINAL_DATA = "/home/hans/data/Machine learning training set (copy)/"
+ORIGINAL_DATA = "/home/hans/data/Machine learning training set/"
 DATA_DIR = "/home/hans/data/bbox_to_segmentation_gt/"
 
-for img in [os.path.join(ORIGINAL_DATA, x) for x in  BBOX_ANNOTATION_FILES]:
+original_images = [os.path.join(ORIGINAL_DATA, x) for x in  BBOX_ANNOTATION_FILES]
+
+for idx, img in enumerate(original_images):
     
+    if not os.path.exists(img):
+        continue
+
+    print ("Working on image - whole body ann pair %d/%d" %(idx+1, len(original_images)))
+    print (img)
     try:
         os.mkdir(os.path.join(DATA_DIR, "original image"))
     except Exception:
@@ -48,6 +40,7 @@ for img in [os.path.join(ORIGINAL_DATA, x) for x in  BBOX_ANNOTATION_FILES]:
     shutil.copyfile(img, os.path.join(DATA_DIR, "original image", img.split('/')[-1]))
 
     ATLEAST_ONE = False
+    annotations = {}
     for directory in glob.glob(os.path.join(os.path.dirname(os.path.dirname(img)), '*')):
         part = directory.split('/')[-1]
         
@@ -59,20 +52,27 @@ for img in [os.path.join(ORIGINAL_DATA, x) for x in  BBOX_ANNOTATION_FILES]:
             pass
    
         try:
-            mask = glob.glob(os.path.join(directory, '*' + img.split('/')[-1].split('.')[0] + '*'))[0]
-            ATLEAST_ONE = True
-            shutil.copyfile(mask, os.path.join(DATA_DIR, part, mask.split('/')[-1]))
-            
+           
+            mask = glob.glob(os.path.join(os.path.dirname(directory), part, '*' + img.split('/')[-1].split('.')[0] + '*'))[0]
+
             if part == "whole body":
                 
-                img_np, mask_np = cv2.imread(img, -1), cv2.imread(mask, -1)
+                if os.path.exists(os.path.join(DATA_DIR, part, mask.split('/')[-1])):
+                    continue
+
+                shutil.copyfile(mask, os.path.join(DATA_DIR, part, mask.split('/')[-1]))
+                
+                img_np, mask_np = cv2.imread(img), cv2.imread(mask)
+                
                 mask_binary = mask_np.copy()
                 mask_binary[mask_binary > 225] = 0
                 mask_binary[mask_binary > 0] = 1
 
                 img_np = cv2.cvtColor(img_np, cv2.COLOR_BGR2GRAY)
                 mask_binary = cv2.cvtColor(mask_binary, cv2.COLOR_BGR2GRAY)
-                mask_np = cv2.cvtColor(mask_np, cv2.COLOR_BGR2GRAY) * mask_binary
+                mask_np = cv2.cvtColor(mask_np, cv2.COLOR_BGR2GRAY)
+                
+                annotations[part] = mask_np 
                 
                 loc = np.zeros((img_np.shape[0]-mask_np.shape[0]+1, img_np.shape[1]-mask_np.shape[1]+1))
                 for idx in range(0, img_np.shape[0]-mask_np.shape[0]+1):
@@ -82,51 +82,33 @@ for img in [os.path.join(ORIGINAL_DATA, x) for x in  BBOX_ANNOTATION_FILES]:
 
                 min_idxs = np.unravel_index(loc.argmin(), loc.shape)
                 
-                cv2.imshow('img', img_np)
-                img_np[min_idxs[0]: min_idxs[0]+mask_np.shape[0], min_idxs[1]:min_idxs[1]+mask_np.shape[1]] = mask_np
-                cv2.imshow('mask', img_np.astype(np.uint8))
-                cv2.waitKey()
+                mask_save = np.zeros_like(img_np)
+                mask_save[min_idxs[0]: min_idxs[0]+mask_np.shape[0], min_idxs[1]:min_idxs[1]+mask_np.shape[1]] = mask_np
+                mask_save[mask_save>225] = 0
+                
+                cv2.imwrite(os.path.join(DATA_DIR, part, mask.split('/')[-1]), mask_save)
+                os.remove(mask)
+
+                ATLEAST_ONE = True
+
+            else:
+                
+                #print ("Saving %s to %s" % (mask, os.path.join(DATA_DIR, part, mask.split('/')[-1])))
+                shutil.copyfile(mask, os.path.join(DATA_DIR, part, mask.split('/')[-1]))
+                os.remove(mask)
 
         except Exception:
             traceback.print_exc()
             pass
 
-    #if not ATLEAST_ONE:
-    #    os.remove(os.path.join(DATA_DIR, "original image", img.split('/')[-1]))
+    os.remove(img)
 
-def create_mask(img_shape, img_file, mask_file, color=WHITE):
-    # Create bounding box based on white background
+# To ensure delete happens after checking the current state of newly created dataset
+exit()
 
-    img = cv2.imread(img_file)
+for fl in original_images:
+    img_name = fl.split('/')[-1].split('.')[0]
+    mask_paths = glob.glob(os.path.join(os.path.dirname(os.path.dirname(fl)), "*", img_name+"*"))
     
-    if img is None:
-        return 
-
-    img = cv2.cvtColor(img, cv2.COLOR_BGR2HSV)
-    
-    mask = cv2.imread(mask_file)
-    mask = cv2.inRange(img, color[0], color[1])
-    #mask = cv2.cvtColor(mask, cv2.COLOR_HSV2BGR)
-    
-    mask = 255 - mask
-    where = np.array(np.where(mask))
-    x1, y1 = np.amin(where, axis=1)
-    x2, y2 = np.amax(where, axis=1)
-    
-    img = cv2.cvtColor(img, cv2.COLOR_HSV2BGR)
-    img = cv2.rectangle(img, (y1, x1), (y2, x2), (255,0,0), 2)
-     
-    #mask = cv2.resize(mask, (y2-y1+1, x2-x1+1))
-    mask_ret = np.zeros_like(img[:,:,:1])
-    #mask_ret[x1:x2+1, y1:y2+1, 0] = mask
-    
-    cv2.imshow('f', img)
-    cv2.imshow('g', mask)
-    cv2.waitKey()
-
-    cv2.imwrite("sample.png", mask)
-    cv2.imwrite("sample_img.png", img)
-
-for data in BBOX_ANNOTATION_FILES:
-    imgpath, annpath, color = data[0], data[1], data[2]
-    create_mask((256,256,3), imgpath, annpath, color)
+    for f in mask_paths:
+        os.remove(f)
